+++ src/app/components/newNoteForm/styles.js
@@ -1,0 +1,65 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  titleInput: {
+    margin: 12,
+    padding: 10,
+    fontSize: 24,
+  },
+  textInput: {
+    // margin: 12,
+    padding: 10,
+    fontSize: 16,
+  },
+  wordCountText: {
+    margin: 5,
+  },
+  centeredView: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: 22,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+  },
+  modalView: {
+    margin: 20,
+    backgroundColor: "white",
+    width: 300,
+    borderRadius: 20,
+    padding: 35,
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+    elevation: 5
+  },
+  button: {
+    borderRadius: 5,
+    padding: 10,
+    elevation: 2,
+    backgroundColor: "#ff55dd",
+    margin: 3,
+  },
+  buttonClose: {
+    borderRadius: 5,
+    padding: 10,
+    elevation: 2,
+    backgroundColor: "red",
+    margin: 3,
+  },
+  textStyle: {
+    color: "white",
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+  modalText: {
+    marginBottom: 15,
+    textAlign: "center"
+  }
+});
+
+export default styles;
